moved spaceContextMenu to context-menus
